requirements.txt
yfinance
pandas
matplotlib
google-auth-oauthlib
google-auth
google-api-python-client
schedule
# dotenv
sendgrid